feat: calculate and display total absence hours and work hours.
